Improve prompt management
